requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
langchain
langchain-ollama
langgraph
httpx
pywin32
